Revert "Handle the change for UnaryPositive now extending RuntimeReplaceable"
This reverts commit 414db71ce732157f26101fcd6e7ec0bb808ebc9b.
